pyproject.toml
[project]
name = "django-page-resolver"
version = "0.1.2"
description = "Utility class for Django that helps determine the page number on which a specific object appears in a paginated list, either from a QuerySet or via a ForeignKey relationship."
authors = [
    {name = "StanislavJPG",email = "samper.stas@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"


[tool.poetry.group.test.dependencies]
factory-boy = "^3.3.3"
